add pic in register form
