+++ src/App.js
@@ -1,16 +1,17 @@
 import './App.css';
 import 'bootstrap/dist/css/bootstrap.css';
 import NavBar from './components/NavBar';
-import Login from './components/auth/Login';
+import Login from './components/auth/LoginInfo';
 import Footer from './components/Footer';
 import MainContent from './components/MainContent';
-import Register from './components/auth/Register';
+import Register from './components/auth/RegisterInfo';
 import 'font-awesome/css/font-awesome.min.css';
 import React, { Suspense, useState} from 'react';
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import TestimonyPage from './components/addTestimonies/TestimonyPage';
 import Book from './components/BookAppointment/Book';
 import Content from './components/BookAppointment/Content';
+import Signup from './components/auth/Signup';
 function App() {
   const [currentForm, setCurrentForm] = useState('register');
   const [homePage, setHomePage] = useState('home')
@@ -36,7 +37,7 @@ function App() {
         <Route exact path="/login" element={<Login/>}/>
       </Routes>
       <Routes>
-        <Route exact path="/register" element={<Register/>}/>
+        <Route exact path="/register" element={<Signup/>}/>
       </Routes>
       <Routes>
         <Route exact path="/testimony" element={<TestimonyPage/>}/>
