+++ src/components/auth/RegisterInfo.js
